requirements.txt
mkdocs-material
mkdocs-bootstrap-tables-plugin